Improve pacman hook
The hook now reacts on updated files instead of updated packages, like
the `mkinitcpio` hook does.

This allows it to be triggered each time a kernel image is updated:
 - when a kernel package is updated (e.g. `linux`);
 - when a dependency is updated (e.g. `systemd`).

This also allows to only rebuild the kernel images that were actually
updated instead of always rebuilding all. The main script now accepts
kernel names as arguments for that.

Files have been renamed to be more consistent with the AUR package.